voyaga/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
openai==1.12.0
